# docs/swagger.yaml
basePath: /
definitions:
...
  models.CreatePermission:
    properties:
      name:
        type: string
    type: object
...
  models.Permission:
    properties:
      created_at:
        type: string
      created_by:
        type: string
      id:
        type: integer
      name:
        type: string
      updated_at:
        type: string
      updated_by:
        type: string
    type: object
...
info:
  contact:
    name: API Support
  description: This is an RBAC based full fledge API serice for managing users, roles
    and permissions in the system
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Truvest Identity Management Service APIs
  version: "1.0"
paths:
...
  /permissions:
    get:
      consumes:
      - application/json
      description: Get all permissions in the system. In order to access this API,
        someone must have "VIEW_PERMISSION" Permission tagged to its role.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Permission'
      security:
      - ApiKeyAuth: []
      summary: Get all Permissions in the system
      tags:
      - Permissions
    post:
      consumes:
      - application/json
      description: Create a permission in the system.User must have "MANAGE_PERMISSION"
        permission tagged to its role in order to access this API
      parameters:
      - description: Create Permission
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.CreatePermission'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Permission'
      security:
      - ApiKeyAuth: []
      summary: Create a permission in the system
      tags:
      - Permission
  /permissions/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a permission by id in the system. In order to access this
        API, someone must have "MANAGE_PERMISSION" Permission tagged to its role.
      parameters:
      - description: id of the Permission
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: ""
      security:
      - ApiKeyAuth: []
      summary: Delete a permission by id in the system
      tags:
      - Permission
    get:
      consumes:
      - application/json
      description: Get details of a permission by ID in the system. In order to access
        this API, someone must have "VIEW_PERMISSION" Permission tagged to its role.
      parameters:
      - description: ID of the permission to get
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Permission'
      security:
      - ApiKeyAuth: []
      summary: Get a Permission by id in the system
      tags:
      - Permission
    put:
      consumes:
      - application/json
      description: Update a permission by id in the system. In order to access this
        API, someone must have "MANAGE_PERMISSION" Permission tagged to its role.
      parameters:
      - description: id of the Permission
        in: path
        name: id
        required: true
        type: integer
      - description: Create Permission
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.CreatePermission'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Permission'
      security:
      - ApiKeyAuth: []
      summary: Update a permission by id in the system
      tags:
      - Permission
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"